+++ src/main/kotlin/com/zephir/ide/highlight/ZephirSyntaxHighlighterFactory.kt
@@ -1,0 +1,16 @@
+// Copyright (c) 2014-2020 Zephir Team
+//
+// This source file is subject the MIT license, that is bundled with
+// this package in the file LICENSE, and is available through the
+// world-wide-web at the following url:
+//
+// https://github.com/zephir-lang/idea-plugin/blob/master/LICENSE
+
+package com.zephir.ide.highlight
+
+import com.intellij.openapi.fileTypes.SingleLazyInstanceSyntaxHighlighterFactory
+import com.intellij.openapi.fileTypes.SyntaxHighlighter
+
+class ZephirSyntaxHighlighterFactory : SingleLazyInstanceSyntaxHighlighterFactory() {
+    override fun createHighlighter(): SyntaxHighlighter = ZephirSyntaxHighlighter()
+}