+++ src/main/kotlin/com/zephir/lang/core/completion/ZephirCompletionContributor.kt
@@ -1,0 +1,21 @@
+// Copyright (c) 2014-2020 Zephir Team
+//
+// This source file is subject the MIT license, that is bundled with
+// this package in the file LICENSE, and is available through the
+// world-wide-web at the following url:
+//
+// https://github.com/zephir-lang/idea-plugin/blob/master/LICENSE
+
+package com.zephir.lang.core.completion
+
+import com.intellij.codeInsight.completion.CompletionContributor
+import com.intellij.codeInsight.completion.CompletionType
+import com.intellij.patterns.PlatformPatterns.psiElement
+import com.zephir.lang.core.ZephirLanguage
+
+// TODO(serghei): re-visit this later
+class ZephirCompletionContributor : CompletionContributor() {
+    init {
+        extend(CompletionType.BASIC, psiElement().withLanguage(ZephirLanguage), ZephirCompletionProvider())
+    }
+}