+++ src/main/kotlin/com/zephir/lang/core/completion/ZephirCompletionSuggestor.kt
@@ -1,0 +1,16 @@
+// Copyright (c) 2014-2020 Zephir Team
+//
+// This source file is subject the MIT license, that is bundled with
+// this package in the file LICENSE, and is available through the
+// world-wide-web at the following url:
+//
+// https://github.com/zephir-lang/idea-plugin/blob/master/LICENSE
+
+package com.zephir.lang.core.completion
+
+import com.intellij.codeInsight.completion.CompletionParameters
+import com.intellij.codeInsight.completion.CompletionResultSet
+
+interface ZephirCompletionSuggestor {
+    fun addCompletions(parameters: CompletionParameters, result: CompletionResultSet)
+}