+++ src/main/kotlin/com/zephir/lang/core/psi/ZephirFile.kt
@@ -1,0 +1,19 @@
+// Copyright (c) 2014-2020 Zephir Team
+//
+// This source file is subject the MIT license, that is bundled with
+// this package in the file LICENSE, and is available through the
+// world-wide-web at the following url:
+//
+// https://github.com/zephir-lang/idea-plugin/blob/master/LICENSE
+
+package com.zephir.lang.core.psi
+
+import com.intellij.psi.FileViewProvider
+import com.intellij.extapi.psi.PsiFileBase
+import com.zephir.lang.core.ZephirFileType
+import com.zephir.lang.core.ZephirLanguage
+
+class ZephirFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, ZephirLanguage) {
+    override fun getFileType() = ZephirFileType
+    override fun toString() = "Zephir File"
+}